requirements.txt
apache-airflow
apache-airflow[google]
pytest
pytest-watcher